fix: /ko structured-data·twitter 메타 한국어 로컬라이즈
* root layout에서 Person/WebSite JSON-LD script 제거
* app/page.tsx(en)에 en.meta.description 기반 Person/WebSite script 추가
* app/ko/page.tsx에 ko.meta.description 기반 Person/WebSite script 추가
* app/ko/layout.tsx에 twitter 메타 한국어 오버라이드 추가
* contents/profile/readme.md body_ko의 'backend' → '백엔드' 수정
