+++ contents/profile/readme.md
@@ -30,7 +30,7 @@ cicd:
   - ArgoCD
   - Jenkins
 body_ko: |
-  backend·AI 시스템을 설계하고 운영한다. RAG 파이프라인, 분산 시스템,
+  백엔드·AI 시스템을 설계하고 운영한다. RAG 파이프라인, 분산 시스템,
   쿠버네티스 GitOps 인프라에 관심이 있으며, 개인 블로그와 서비스를
   직접 개발·운영하면서 기록을 남긴다.
 body_en: |
